Exclusion part now works except for Uranium items. ElementGroup now creates Chunk item and Ore block. Items are now sorted in a logical way in Resources Tab. Updated lang files.

+++ src/main/java/net/intercraft/intercraftcore/item/group/ItemUraniumGroup.java
@@ -12,7 +12,7 @@ public class ItemUraniumGroup extends ItemElementGroup
     }
 
     @Override
-    protected Item createItem(byte form, String suffix)
+    protected Item createItem(int form, String suffix)
     {
         return new ItemUranium(element, suffix);
     }
